highlight shader tests
Old attempts at highlight shader were a failure, and going through all tutorials and documentation show it may not be even possible with 3d objects with shadergraph. Added a couple of tests, and set one shader that basically just creates a blue and slightly larger version of the object to the test socketables. Much more tweaking to be had, but this is a thing to do while working on other things. Not a priority