Textures and UI Updates
Added the 3 missing inventory icons to the HUD. Moved textures into the Materials\Texture Folder. Applied textures to all relevant models. Created 2 prefabs for the Target Indicator Model, one for Fish and one for Harvestables. Both of the prefabs, plus the model, for the Target Indicator are in the UI Folder.